requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai==0.27.8
openai-whisper==20230314
python-dotenv==1.0.0
pytube @ git+https://github.com/oncename/pytube.git@6c45936b9703ce986ccb8d0d3595c7974716f94b
regex==2023.6.3
requests==2.31.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.0
urllib3==2.0.3
yarl==1.9.2
youtube-transcript-api==0.6.1
